requirements.txt
pandas
numpy
pymysql
python-dotenv
matplotlib
scikit-learn
